Add files via upload
This is an updated schematic for the ESP32 based module for the IC-705 to XPA125B remote band switching.  This version adds transmit key isolation, a 12v to 5v power supply inspired by Tilman SM0JZT to eliminate the ESP32 development board, and a .5 amp fuse for XPA125B protection.  I have tested this circuit as drawn and it has worked perfectly.